1) Modified the query in the 'getDatasets route'. Now it also creates a 'type' field for the dataset nodes, which is needed to create a special class specific to dataset nodes (i.e. different sizes/color).
2) Modified DFService so it stopped using both dtarget and dsource to
only use dsource (dtarget was not necessary). It was initially creating
an edge with [pTarget,dTarget] = [-1,0] which cannon't be done. This was
fixed quite easily by using dSource which got its value from getting the
index of the dataSet node during each iteration in the while loop. Also
assigning the value of pTarget to the counter i needed to be done before
creating the relation.

Currently when you check a box on the webapp [CAN ONLY CHECK 1 at a
time] The dataset that is chosen is interpolated in the webapp. However
if you submit the same or a different dataset, the process repeats, even
for the same dataset, and removes the force for the previous dataset
interpolated.