this block was added to the pom.xml dependencies:         <dependency> 			<groupId>javax.servlet</groupId> 			<artifactId>javax.servlet-api</artifactId> 			<version>3.0.1</version> 		    <scope>provided</scope> 		</dependency>
After adding this, updating the project and redeploying, my project
started working.  But if I went back and removed this block and
redeployed, the project still works, so I am not sure if it truly helps.

Other notes:

The first line of frontend.js now has a variable that holds a string
that should be identical to the name of the war file deployed to tomcat.
It is easier than finding all the lines and changing them.

You will have to change the port number in Util.java to reflect your
configuration.